# docs/openapi/v1.yaml
openapi: 3.1.0
info:
  title: HamoPMS API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
  description: |
    Versioned contract for the Sprint 1 HamoPMS endpoints. All successful JSON
    responses use a `data` envelope and errors use the shared `error` envelope.
servers:
  - url: /api/v1
...
paths:
...
  /auth/reset-password:
    post:
      tags: [Authentication]
      operationId: resetPassword
      summary: Reset a password with a valid broker token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset and existing API tokens revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum token
...
  responses:
...
    ValidationError:
      description: The request contains invalid data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ResetPasswordRequest:
      type: object
      additionalProperties: false
      required: [email, token, password, password_confirmation]
      properties:
        email:
          type: string
          format: email
        token:
          type: string
        password:
          type: string
          minLength: 12
          format: password
          description: Must contain upper- and lowercase letters, a number, and a symbol.
        password_confirmation:
          type: string
          minLength: 12
          format: password
...
    MessageResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [message]
          properties:
            message:
              type: string
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples: [VALIDATION_FAILED]
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Error'
    ValidationErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              const: VALIDATION_FAILED
            message:
              type: string
            details:
              type: object
              required: [fields]
              properties:
                fields:
                  type: object
                  additionalProperties:
                    type: array
                    items:
                      type: string